requirements.txt
Flask==2.2.5
ipython==8.24.0
langchain_core==0.3.0
langchain_openai==0.2.0
python-dotenv==1.0.1
gunicorn
caldav